Removed container component
