+++ src/main/java/org/liquidengine/legui/marshal/json/gsonimpl/GsonMarshalRegistry.java
@@ -11,7 +11,6 @@ import org.liquidengine.legui.border.SimpleLineBorder;
 import org.liquidengine.legui.component.Button;
 import org.liquidengine.legui.component.CheckBox;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Frame;
 import org.liquidengine.legui.component.ImageView;
 import org.liquidengine.legui.component.Label;
@@ -45,7 +44,6 @@ import org.liquidengine.legui.marshal.json.gsonimpl.border.GsonSimpleLineBorderM
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonButtonMarshaller;
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonCheckBoxMarshaller;
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonComponentMarshaller;
-import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonContainerMarshaller;
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonFrameMarshaller;
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonImageViewMarshaller;
 import org.liquidengine.legui.marshal.json.gsonimpl.component.GsonLabelMarshaller;
@@ -210,7 +208,6 @@ public class GsonMarshalRegistry implements JsonMarshalRegistry {
             // components
             I.registerMarshaller("Button", Button.class, new GsonButtonMarshaller<>());
             I.registerMarshaller("CheckBox", CheckBox.class, new GsonCheckBoxMarshaller<>());
-            I.registerMarshaller("Container", Container.class, new GsonContainerMarshaller<>());
             I.registerMarshaller("Component", Component.class, new GsonComponentMarshaller<>());
             I.registerMarshaller("ImageView", ImageView.class, new GsonImageViewMarshaller<>());
             I.registerMarshaller("Label", Label.class, new GsonLabelMarshaller<>());
