+++ src/main/java/org/liquidengine/legui/system/handler/WindowFocusEventHandler.java
@@ -2,7 +2,6 @@ package org.liquidengine.legui.system.handler;
 
 import java.util.List;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Frame;
 import org.liquidengine.legui.component.Layer;
 import org.liquidengine.legui.event.WindowFocusEvent;
@@ -27,11 +26,10 @@ public class WindowFocusEventHandler extends AbstractSystemEventHandler<SystemWi
             return;
         }
         EventProcessor.getInstance().pushEvent(new WindowFocusEvent(component, context, frame, event.focused));
-        if (component instanceof Container) {
-            List<Component> childs = ((Container) component).getChilds();
-            for (Component child : childs) {
-                pushEvent(child, event, context, frame);
-            }
+
+        List<Component> childs = component.getChilds();
+        for (Component child : childs) {
+            pushEvent(child, event, context, frame);
         }
 
     }
