+++ src/main/java/org/liquidengine/legui/system/handler/WindowIconifyEventHandler.java
@@ -2,7 +2,6 @@ package org.liquidengine.legui.system.handler;
 
 import java.util.List;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Frame;
 import org.liquidengine.legui.component.Layer;
 import org.liquidengine.legui.event.WindowIconifyEvent;
@@ -27,11 +26,9 @@ public class WindowIconifyEventHandler extends AbstractSystemEventHandler<System
             return;
         }
         EventProcessor.getInstance().pushEvent(new WindowIconifyEvent(component, context, frame, event.iconified));
-        if (component instanceof Container) {
-            List<Component> childs = ((Container) component).getChilds();
-            for (Component child : childs) {
-                pushEvent(child, event, context, frame);
-            }
+        List<Component> childs = component.getChilds();
+        for (Component child : childs) {
+            pushEvent(child, event, context, frame);
         }
     }
 }
