+++ src/main/java/org/liquidengine/legui/system/handler/WindowRefreshEventHandler.java
@@ -2,7 +2,6 @@ package org.liquidengine.legui.system.handler;
 
 import java.util.List;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Frame;
 import org.liquidengine.legui.component.Layer;
 import org.liquidengine.legui.event.WindowRefreshEvent;
@@ -26,11 +25,9 @@ public class WindowRefreshEventHandler extends AbstractSystemEventHandler<System
             return;
         }
         EventProcessor.getInstance().pushEvent(new WindowRefreshEvent(component, context, frame));
-        if (component instanceof Container) {
-            List<Component> childs = ((Container) component).getChilds();
-            for (Component child : childs) {
-                pushEvent(child, context, frame);
-            }
+        List<Component> childs = component.getChilds();
+        for (Component child : childs) {
+            pushEvent(child, context, frame);
         }
     }
 }
