+++ src/main/java/org/liquidengine/legui/system/handler/WindowSizeEventHandler.java
@@ -3,7 +3,6 @@ package org.liquidengine.legui.system.handler;
 import java.util.Collections;
 import java.util.List;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Frame;
 import org.liquidengine.legui.component.Layer;
 import org.liquidengine.legui.event.WindowSizeEvent;
@@ -35,11 +34,9 @@ public class WindowSizeEventHandler implements SystemEventHandler<SystemWindowSi
             return;
         }
         EventProcessor.getInstance().pushEvent(new WindowSizeEvent(component, context, frame, event.width, event.height));
-        if (component instanceof Container) {
-            List<Component> childs = ((Container) component).getChilds();
-            for (Component child : childs) {
-                pushEvent(child, event, context, frame);
-            }
+        List<Component> childs = component.getChilds();
+        for (Component child : childs) {
+            pushEvent(child, event, context, frame);
         }
     }
 }
