+++ src/main/java/org/liquidengine/legui/theme/dark/DarkTheme.java
@@ -3,7 +3,6 @@ package org.liquidengine.legui.theme.dark;
 import org.liquidengine.legui.component.Button;
 import org.liquidengine.legui.component.CheckBox;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Label;
 import org.liquidengine.legui.component.LayerContainer;
 import org.liquidengine.legui.component.ProgressBar;
@@ -23,7 +22,6 @@ import org.liquidengine.legui.theme.ThemeManager;
 import org.liquidengine.legui.theme.dark.def.DarkButtonTheme;
 import org.liquidengine.legui.theme.dark.def.DarkCheckBoxTheme;
 import org.liquidengine.legui.theme.dark.def.DarkComponentTheme;
-import org.liquidengine.legui.theme.dark.def.DarkContainerTheme;
 import org.liquidengine.legui.theme.dark.def.DarkLabelTheme;
 import org.liquidengine.legui.theme.dark.def.DarkLayerContainerTheme;
 import org.liquidengine.legui.theme.dark.def.DarkProgressBarTheme;
@@ -62,7 +60,6 @@ public class DarkTheme extends Theme {
         manager.setComponentTheme(Button.class, new DarkButtonTheme<>());
         manager.setComponentTheme(CheckBox.class, new DarkCheckBoxTheme<>());
         manager.setComponentTheme(Component.class, new DarkComponentTheme<>());
-        manager.setComponentTheme(Container.class, new DarkContainerTheme<>());
         manager.setComponentTheme(Label.class, new DarkLabelTheme<>());
         manager.setComponentTheme(LayerContainer.class, new DarkLayerContainerTheme<>());
         manager.setComponentTheme(ProgressBar.class, new DarkProgressBarTheme<>());
