+++ src/main/java/org/liquidengine/legui/theme/dark/def/DarkLayerContainerTheme.java
@@ -8,7 +8,7 @@ import org.liquidengine.legui.component.LayerContainer;
  *
  * @param <T> {@link LayerContainer} subclasses.
  */
-public class DarkLayerContainerTheme<T extends LayerContainer> extends DarkContainerTheme<T> {
+public class DarkLayerContainerTheme<T extends LayerContainer> extends DarkComponentTheme<T> {
 
     /**
      * Used to apply theme only for component and not apply for child components.
