+++ src/main/java/org/liquidengine/legui/theme/dark/def/DarkScrollablePanelTheme.java
@@ -1,7 +1,7 @@
 package org.liquidengine.legui.theme.dark.def;
 
 import org.liquidengine.legui.color.ColorConstants;
-import org.liquidengine.legui.component.Container;
+import org.liquidengine.legui.component.Component;
 import org.liquidengine.legui.component.ScrollablePanel;
 import org.liquidengine.legui.theme.Themes;
 
@@ -20,7 +20,7 @@ public class DarkScrollablePanelTheme<T extends ScrollablePanel> extends DarkCom
     @Override
     public void apply(T component) {
         super.apply(component);
-        Container viewport = component.getViewport();
+        Component viewport = component.getViewport();
         Themes.getDefaultTheme().apply(viewport);
         Themes.getDefaultTheme().applyAll(component.getVerticalScrollBar());
         Themes.getDefaultTheme().applyAll(component.getHorizontalScrollBar());
@@ -36,7 +36,7 @@ public class DarkScrollablePanelTheme<T extends ScrollablePanel> extends DarkCom
     @Override
     public void applyAll(T component) {
         super.applyAll(component);
-        Container viewport = component.getViewport();
+        Component viewport = component.getViewport();
         Themes.getDefaultTheme().applyAll(viewport);
         viewport.setBorder(null);
         viewport.setBackgroundColor(ColorConstants.transparent());
