+++ src/main/java/org/liquidengine/legui/theme/white/WhiteTheme.java
@@ -3,7 +3,6 @@ package org.liquidengine.legui.theme.white;
 import org.liquidengine.legui.component.Button;
 import org.liquidengine.legui.component.CheckBox;
 import org.liquidengine.legui.component.Component;
-import org.liquidengine.legui.component.Container;
 import org.liquidengine.legui.component.Label;
 import org.liquidengine.legui.component.LayerContainer;
 import org.liquidengine.legui.component.ProgressBar;
@@ -23,7 +22,6 @@ import org.liquidengine.legui.theme.ThemeManager;
 import org.liquidengine.legui.theme.white.def.WhiteButtonTheme;
 import org.liquidengine.legui.theme.white.def.WhiteCheckBoxTheme;
 import org.liquidengine.legui.theme.white.def.WhiteComponentTheme;
-import org.liquidengine.legui.theme.white.def.WhiteContainerTheme;
 import org.liquidengine.legui.theme.white.def.WhiteLabelTheme;
 import org.liquidengine.legui.theme.white.def.WhiteLayerContainerTheme;
 import org.liquidengine.legui.theme.white.def.WhiteProgressBarTheme;
@@ -54,7 +52,6 @@ public class WhiteTheme extends Theme {
         manager.setComponentTheme(Button.class, new WhiteButtonTheme<>());
         manager.setComponentTheme(CheckBox.class, new WhiteCheckBoxTheme<>());
         manager.setComponentTheme(Component.class, new WhiteComponentTheme<>());
-        manager.setComponentTheme(Container.class, new WhiteContainerTheme<>());
         manager.setComponentTheme(Label.class, new WhiteLabelTheme<>());
         manager.setComponentTheme(LayerContainer.class, new WhiteLayerContainerTheme<>());
         manager.setComponentTheme(ProgressBar.class, new WhiteProgressBarTheme<>());
