+++ src/main/java/org/liquidengine/legui/theme/white/def/WhiteScrollablePanelTheme.java
@@ -1,7 +1,7 @@
 package org.liquidengine.legui.theme.white.def;
 
 import org.liquidengine.legui.color.ColorConstants;
-import org.liquidengine.legui.component.Container;
+import org.liquidengine.legui.component.Component;
 import org.liquidengine.legui.component.ScrollablePanel;
 import org.liquidengine.legui.theme.Themes;
 
@@ -15,7 +15,7 @@ public class WhiteScrollablePanelTheme<T extends ScrollablePanel> extends WhiteC
     @Override
     public void apply(T component) {
         super.apply(component);
-        Container viewport = component.getViewport();
+        Component viewport = component.getViewport();
         Themes.getDefaultTheme().apply(viewport);
         Themes.getDefaultTheme().apply(component.getVerticalScrollBar());
         Themes.getDefaultTheme().apply(component.getHorizontalScrollBar());
@@ -26,7 +26,7 @@ public class WhiteScrollablePanelTheme<T extends ScrollablePanel> extends WhiteC
     @Override
     public void applyAll(T component) {
         super.applyAll(component);
-        Container viewport = component.getViewport();
+        Component viewport = component.getViewport();
         Themes.getDefaultTheme().applyAll(viewport);
         viewport.setBorder(null);
         viewport.setBackgroundColor(ColorConstants.transparent());
